+++ src/components/ExploreComponents/DealsCard.js
@@ -11,6 +11,9 @@ import { LinearGradient } from "expo-linear-gradient";
 
 const DealsCard = () => {
   return (
+    // <View style={styles.viewContainer}>
+    //   <Text style={styles.dealsText}>Deals</Text>
+
     <ScrollView style={styles.scrollContainer} horizontal={true}>
       {/* <TouchableOpacity style={styles.card}>*/}
       <TouchableOpacity style={styles.container}>
@@ -93,14 +96,14 @@ const styles = StyleSheet.create({
     shadowRadius: 3.84,
     elevation: 5,
     marginHorizontal: 8,
-    width: 100, // Set the width here instead of maxWidth for more control
-    height: 160, // You can set a fixed height to avoid overflowing
+    width: 170, // Set the width here instead of maxWidth for more control
+    height: 230, // You can set a fixed height to avoid overflowing
   },
   dealsText: {
     // alignSelf: "stretch",
     marginTop: 40,
     marginLeft: 20,
-    fontSize: 10,
+    fontSize: 30,
     fontWeight: "bold",
     // marginBottom: 10,
   },
@@ -118,10 +121,9 @@ const styles = StyleSheet.create({
     borderRadius: 18, // Make sure the gradient has rounded corners
   },
   imageContainer: {
-    height: 100,
+    height: 150,
     width: "100%",
     backgroundColor: "#F2F3F2",
-    alignSelf: "center",
   },
   image: {
     width: "100%",
@@ -141,7 +143,7 @@ const styles = StyleSheet.create({
   price: {
     alignSelf: "center",
     fontFamily: "Lato",
-    fontSize: 13,
+    fontSize: 15,
     fontWeight: "700",
     color: "#fff", // Change text color to white for better contrast against gradient
   },
