+++ src/navigation/ExploreNavigator.js
@@ -1,25 +1,14 @@
-import { StyleSheet, Text, View } from "react-native";
-import { createNativeStackNavigator } from "@react-navigation/native-stack";
-import React from "react";
-import ProductDetailsScreen from "../screens/shopping/ProductDetailsScreen";
-import ExploreScreen from "../screens/shopping/ExploreScreen";
-import ShopDetailsScreen from "../screens/shopping/ShopDetailsScreen";
-import ShopLinkScreen from "../screens/shopping/ShopLinkScreen";
+// import { StyleSheet, Text, View } from 'react-native'
+// import React from 'react'
 
-const Stack = createNativeStackNavigator();
-const ShoppingNavigation = () => {
-  return (
-    <Stack.Navigator
-      screenOptions={{
-        headerShown: false,
-      }}
-    >
-      <Stack.Screen name="ExploreScreen" component={ExploreScreen} />
-      <Stack.Screen name="ProductDetails" component={ProductDetailsScreen} />
-    </Stack.Navigator>
-  );
-};
+// const ExploreNavigator = () => {
+//   return (
+//     <View>
+//       <Text>ExploreNavigator</Text>
+//     </View>
+//   )
+// }
 
-export default ShoppingNavigation;
+// export default ExploreNavigator
 
-const styles = StyleSheet.create({});
+// const styles = StyleSheet.create({})
